Build Comtrade Data from Temporary Files

Builds final Comtrade data set from temporary files downloaded using download_Comtrade(..., rm.temporaryFiles = FALSE).

Options rm.temporaryFiles = FALSE in download_Comtrade() are necessary as temporary files will be deleted as part of data download by default. See documentation of download_Comtrade() for more information.

build_Comtrade(
  directory = system.file("data", package = "comtradeRggregator"),
  rm.temporaryFiles = FALSE,
  is.mirrorData = FALSE,
  partner = NULL
)

Arguments

directory
Location of temporary file downloads; default is

<your local package directory>\data\tmp\<date-and-time-stamp>.

rm.temporaryFiles
Remove temporary download files stored at location provided in argument location.temporaryFiles when building Comtrade trade data set; default is TRUE.

is.mirrorData
Is data set that is to be built mirror data? default is FALSE, i.e. non-mirror data.

partner
List of partner countries. Needs to be provided if is.mirrorData = TRUE; default is NULL, i.e. no mirror data download.

If is.mirrorData = TRUE, export (import) data from countries specified in argument partner is measured as import (export) of countries specified in argument partner as reported by countries specified in argument partner.

See also