Check Comtrade token

Checks validity of supplied Comtrade token as described in Comtrade's API documentation.

The validity of the token can be checked using the official Access Rights Information mask, or directly via the official API service (link).

In order to add a valid token to your comtradeRggregator query:

Follow the instruction on Comtrade's API page to Comtrade's API help page to get access to an authentication code (token).

You can check the validity of your token here.

Use the token as input parameter token in function download_Comtrade() to get unlimited (authenticated) usage of 10,000 requests per hour. You can check the validity of the token obtained in steps (1) to (2) without running a comtradeRggreagator query by using the function check_token(token = "<your token>".

check_token(token = NULL)

Arguments

token
Comtrade token; default is NULL.

Examples

check_token() # by default, no token provided.

#> No Comtrade token specified; download restricted to 100 queries per hour.

if (FALSE) {
check_token(token = "your token") # add your access token
}