Convert Comtrade Trade Data

Offers conversion tables for Comtrade trade data following official tables from the UN Statistics Division (UNSD) and the World Integrated Trade Solution (WITS). For a complete list of concordance tables please see comtradeRggregator's Concordance Table. Please see here for a list of all nomenclatures supported by comtradeRggregator.

convert_Comtrade(
  data,
  classification.from = "H3",
  commodity.code = "commodity_code",
  classification.to = "I3"
)

Arguments

data
A data frame or vector containing (a variable listing) commodity codes following one of the supported nomenclatures; see here.

classification.from
Abbreviation of origin classification supported by comtradeRggregator's Concordance Table.

commodity.code
Name of variable containing the commodity codes corresponding to trade classification; default is commodity_code.

classification.to
Abbreviation of target classification based on comtradeRggregator's Concordance Table.

Source

https://unstats.un.org/unsd/classifications/Econ#corresp-hs https://wits.worldbank.org/product_concordance.html https://unstats.un.org/unsd/classifications/Econ#corresp-hs https://amannj.github.io/resources/comtradeRggregator/index.html#trade-classifications https://amannj.github.io/resources/comtradeRggregator/index.html#concordance-table

Examples

df <- tibble("H3 commodity code" = c("010110", "071040"))
convert_Comtrade(df,
  classification.from = "H3",
  commodity.code = "H3 commodity code",
  classification.to = "I3"
)

#> # A tibble: 2 x 4
#>   `H3 commodity co~ `HS 2007 Product De~ `ISIC Revision 3~ `ISIC Revision 3 Pro~
#>   <chr>             <chr>                <chr>             <chr>                
#> 1 010110            Live horses/asses/m~ 0121              Farming of cattle, s~
#> 2 071040            Sweet corn, uncooke~ 1513              Processing and prese~

if (FALSE) {
# Extract Comtrade trade data first
AT_World <- download_Comtrade(
  year = "2018",
  frequency = "annual",
  countries = "Austria",
  partners = "World",
  tradecode = "HS2007",
  ag = "AG6",
  type = "commodities",
  select.stats = "trade_value_usd",
  direction = "all"
)

# Convert HS2007/H3 trade data to ISIC Rev. 3
AT_World %>%
  convert_Comtrade(
    classification.from = "H3",
    commodity.code = "commodity_code",
    classification.to = "I3"
  )
}