Extract Comtrade Data

Main function to generate trade data query that is processed in download_Comtrade_core() to interact with comtradr::ct_search() and Comtrade's API to extract Comtrade trade data.

download_Comtrade(
  year = "2018",
  frequency = "annual",
  month = NULL,
  countries = "all",
  partners = "World",
  nomenclature = "HS2007",
  commodity = "AG6",
  type = "commodities",
  select.stats = "all",
  direction = "all",
  token = NULL,
  ext_cnt = 5,
  is.mirrorData = FALSE,
  rm.temporaryFiles = TRUE,
  location.temporaryFiles = NULL,
  build.Comtrade = TRUE,
  sleep = 5
)

Arguments

year
Year for which to extract data.

frequency
Frequency of data extract; either annual or monthly; default is annual.

month
Optional parameter for monthly extract; ignored for annual extracts; default is all 12 months if frequency = 'monthly' while argument monthy remains unspecified.

countries
Select list of countries to be extracted; default is all.

partners
Specify partner country/countries or World (as provided by Comtrade); use option all if you want to download trade data from all available countries but not Comtrade's World aggregate; default is World.

If is.mirrorData = FALSE and partners = 'World', download_Comtrade() will download aggregated "World" trade as provided by Comtrade.

If is.mirrorData = TRUE and partners = 'World', no "World" aggregate is available via Comtrade. In this case, download_Comtrade() will download mirror trade data for all available countries, and creates and returns an artificial 'World' aggregate by summing up over all available countries.

nomenclature
Select nomenclature and corresponding trade database for data extraction; default is HS2007/H3; monthly trade data only available for HS. The full list of possible trade classifications and corresponding input arguments are provided in table Trade Classification (link); argument accepts long classification names, e.g. HS2007, as well as abbreviations, e.g. HS3, as inputs.

commodity
Specify commodity code or trade data aggregate for download. Both vary by trade data set; see Trade Classification (link) for more information.

type
Type of trade data to be extracted (either services or commodities); currently only type = commodities implemented.

select.stats
Trade statistics to be reported; either trade_value_usd, qty_unit_code, qty_unit, alt_qty_unit_code, alt_qty_unit, qty, alt_qty, netweight_kg, gross_weight_kg or all; default is all.

direction
Direction of trade flow reported; either imports, exports, re-imports, re-exports or all; default is all.

token
Set Comtrade token to increase hourly queries from Comtrade API; increases the speed of large data extraction as it increases the number of queries that can be run per hour.

ext_cnt
Number of countries extracted with each query; default is 5 which is also the maximum in comtradr. In case your query exceeds the maximum resultset allowed by the comtrade API, reduce the number of countries you extract with every query using the ext_cnt parameter. For example, if a data extraction request fails with the default parameter value ext_cnt = 5, try reducing it to, say, ext_cnt = 3. This reduces the size of the resultset.

is.mirrorData
Extract mirror trade data? default is FALSE; set to TRUE to extract mirror trade data from country/countries specified in argument partners. For example, if is.mirrorData = TRUE export data from countries specified in argument countries to countries specified in argument partners is measured as import data from countries specified in argument countries to countries specified in argument partners as reported by countries specified in argument partners.

rm.temporaryFiles
Remove temporary download files stored at location provided in argument location.temporaryFiles; default is TRUE.

location.temporaryFiles
Location of temporary file downloads; default is your local package directory\data\tmp\<date-and-time-stamp>; if you decide to use a different location, an alternative temporary folder needs to be created first.

build.Comtrade
After having downloaded the separate temporary files (which will be saved to location.temporaryFiles), should function download_Comtrade() return a complete data object? Default is TRUE.

sleep
Number of seconds to wait before the next Comtrade API query is started; default is 5.

Examples

if (FALSE) {
AT_World <- download_Comtrade(
  year = "2018",
  frequency = "annual",
  countries = "Austria",
  partners = "World",
  nomenclature = "HS2007",
  commodity = "AG6",
  type = "commodities",
  select.stats = "trade_value_usd",
  direction = "all"
)
AT_World
}