Core function of download_Comtrade()

Interacts with comtradr::ct_search() and Comtrade's API to extract Comtrade trade data.

download_Comtrade_wrapper(
  rep = "Austria",
  date = "2019",
  dir = "all",
  is.mirrorData = FALSE,
  aggregation_level = "AG6",
  select.stats = .select.stats,
  frequency = "annual",
  partners = partners,
  sleep = sleep
)

Arguments

rep
Country/countries to be extracted; default is all.

date
Year/year-month for which to extract data.

dir
Direction of trade flow reported; either imports, exports, re_-_imports, re_-_exports or all; default is all.

is.mirrorData
Extract mirror trade data? default is FALSE; Set to TRUE to extract mirror trade data from country/countries specified in argument partners. For example, if is.mirrorData = TRUE export data from countries specified in argument countries to countries specified in argument partners is measured as import data from countries specified in argument countries to countries specified in argument partners as reported by countries specified in argument partners.

aggregation_level
Level of aggregation of trade data; varies by trade data set.

select.stats
Trade statistics to be reported; either trade_value_usd, qty_unit_code, qty_unit, alt_qty_unit_code, alt_qty_unit, qty, alt_qty, netweight_kg, gross_weight_kg or all; default is all.

frequency
Frequency of data extract; either annual or monthly; default is annual.

partners
Specify partner country/countries or World (as provided by Comtrade) for global, aggregated trade; default is World.

sleep
Number of seconds to wait before the next Comtrade API query is started; default is 20.