Check Available Countries for Comtrade Query

Extract or validate list of exporters for which in a given period (year or year-month) and trade classification data is available. Data availability varies by country and classification, and new trade data is released/revised very frequently. comtradeRggregator uses trade information corresponding to official Comtrade data availability which it updates once a day and stores locally on your machine: your local package directory/data/Comtrade_DataAvailability-<date stamp>.rds The locally stored Comtrade Data Availability file is used in every download_Comtrade() data query. In other words, if new data becomes available and is documented in the official Comtrade data availability file, comtradeRggregator will use this information no later than 24 hours after its publication.

gen_CountryList(
  directory = system.file("data", package = "comtradeRggregator"),
  file = paste0("Comtrade_DataAvailability-", Sys.Date()),
  type = "commodities",
  frequency = "annual",
  time = 2010,
  .px = "ag6",
  time_count = 1
)

Arguments

directory
Location of Comtrade Data Availability file; default is data/

file
Name of file containing information on Comtrade's data availability; default is Comtrade_DataAvailability-<date stamp>.

type
Type of trade data to be extracted (either services or commodities); currently only type = commodities implemented.

frequency
Frequency of data extract; either annual or monthly; default is annual.

time
Generic time id for internal processing; either a particular year or particular month of a particular year

.px
Trade classification abbreviation .px for further processing of Comtrade download

time_count
Generic time index for monthly time processing.