Check Data Availability for Comtrade Trade Data Query

Check if data of a particular trade data set, frequency or country is available. Data availability varies by country and classification, and new trade data is released/revised very frequently on Comtrade. comtradeRggregator uses trade information corresponding to official Comtrade data availability which it updates once a day and stores locally on your machine:

<your local package location>/data/Comtrade_DataAvailability-<date stamp>.csv.gz.

The locally stored Comtrade Data Availability file is used in every download_Comtrade() data query. In other words, if new data becomes available and is documented in the official Comtrade data availability file, comtradeRggregator will use this information no later than 24 hours after its publication.

is.available_Comtrade(
  is.contained = NULL,
  is.fuzzy = FALSE,
  type = "commodities",
  frequency = "annual",
  month = NULL,
  tradecode = "H3",
  year = 2008,
  directory = system.file("data", package = "comtradeRggregator"),
  file = paste0("Comtrade_DataAvailability-", Sys.Date())
)

Arguments

is.contained
Check if a particular country (set of countries) is available for a particular data set; default is NULL which returns list of all countries available for a given data set.

is.fuzzy
Fuzzy match of country name fragments provided in is.contained; default is FALSE.

type
Type of trade data to be extracted; either services or commodities; currently only commodities implemented.

frequency
Frequency of data extract; either annual or monthly; default is annual.

month
Optional parameter for monthly extract; ignored for annual extracts; only takes one entry as monthly trade data availability varies by month.

tradecode
Select trade database and classification to be extracted; default is HS2007; monthly trade data only available following HS classification; the full list of possible trade classifications and their corresponding input arguments used in the comtradeRggregator package are provided in table Trade Classifications.

year
Year for which to extract data.

directory
Location of directory; default is "<location of package on your system>\data".

file
Location of Comtrade Data Availability file; default is Comtrade_DataAvailability-<time and date stamp>.csv.gz.

Examples

if (FALSE) {
is.available_Comtrade(frequency = "annual", tradecode = "H3", year = 2008)
is.available_Comtrade(
  is.contained = c("Austria", "Germany"),
  frequency = "annual", tradecode = "H2", year = 2012
)
is.available_Comtrade(
  is.contained = c("Austria", "Germany"),
  frequency = "monthly", month = "01", tradecode = "H3", year = 2020
)
}